Refactor into core, cli, and app with a `Suitor` facade in prep for interface buildout
The `Suitor::Core` will be shared across two interfaces: the CLI and the Browser. This is where the work horses for the library reside.

For now, `Suitor` really only has one purpose: charm a phone number with a message generated from reddit. Convenient access to this feature provided via the `Suitor.charm` facade.

Suitor also depends on two external services "Twilio" and "Reddit". These services are wrapped so the application boundaries are easier to test and clearly established.

Also leverage the magic numbers for Twilio's test api so we can integration test against it without worrying about money or api limits.